Create Guest Network Password and Number of Connected Devices for Network node
As promised in #479

This PR gives an example how to display the Password of the Guest Network on a Google Nest Hub.
It will not give a response on a Google Mini speaker, as these are not setup to respond to this request.

This PR gives also an example how to configure "Connected devices"
It will respond to a request on both a Google Mini speaker, as well on a Google Nest Hub. It will not been shown on the display.